Place dataset and loss tests separately
